Rename and move Dark Matter tool classes
