+++ src/main/java/com/pahimar/ee3/item/tool/ItemShovelDarkMatter.java
@@ -1,42 +1,48 @@
 package com.pahimar.ee3.item.tool;
 
+
 import net.minecraft.block.Block;
 import net.minecraft.block.material.Material;
 import net.minecraft.client.renderer.texture.IIconRegister;
 import net.minecraft.entity.EntityLivingBase;
 import net.minecraft.entity.player.EntityPlayer;
-import net.minecraft.item.ItemAxe;
+import net.minecraft.item.ItemSpade;
 import net.minecraft.item.ItemStack;
 import net.minecraft.world.World;
 
+import com.pahimar.ee3.client.util.PowerItemUtils;
+import com.pahimar.ee3.creativetab.CreativeTab;
+import com.pahimar.ee3.item.IKeyBound;
+import com.pahimar.ee3.item.ItemEE;
 import com.pahimar.ee3.reference.IChargeable;
 import com.pahimar.ee3.reference.Key;
 import com.pahimar.ee3.reference.Names;
 import com.pahimar.ee3.reference.Textures;
-import com.pahimar.ee3.creativetab.CreativeTab;
-import com.pahimar.ee3.item.IKeyBound;
-import com.pahimar.ee3.item.ItemEE;
 
 import cpw.mods.fml.relauncher.Side;
 import cpw.mods.fml.relauncher.SideOnly;
 
-//[RANT] If Java had multiple inheritance, I wouldn't have to copy-paste
-//the dark-matter-izing override code between DM tools. [/RANT]
 
-public class DMAxe extends ItemAxe implements IChargeable, IKeyBound
+public class ItemShovelDarkMatter extends ItemSpade implements IChargeable, IKeyBound
 {
-
-	public DMAxe()
+	public ItemShovelDarkMatter()
 	{
-		super(DMUtils.MATERIALDARKMATTER);
+		super(PowerItemUtils.MATERIALDARKMATTER);
 		
-		setUnlocalizedName(Names.Tools.AXE_DARK_MATTER);
+		setUnlocalizedName(Names.Tools.SHOVEL_DARK_MATTER);
 		setCreativeTab(CreativeTab.EE3_TAB);
 		
 		setNoRepair();
 		
 		maxStackSize = 1;
 	}
+	
+	//not repairable... because it never breaks
+	@Override
+    public boolean getIsRepairable(ItemStack par1ItemStack, ItemStack par2ItemStack)
+    {
+        return false;
+    }
 	
 	//do not decrease durability on block destroyed
 	@Override
@@ -51,13 +57,13 @@ public class DMAxe extends ItemAxe implements IChargeable, IKeyBound
 	public float getDigSpeed(ItemStack itemstack, Block block, int metadata)
     {
 		
-        if(block.getMaterial() != Material.wood && block.getMaterial() != Material.plants && block.getMaterial() != Material.vine)
+        if(block.getMaterial() != Material.clay && block.getMaterial() != Material.craftedSnow && block.getMaterial() != Material.grass && block.getMaterial() != Material.ground && block.getMaterial() != Material.sand && block.getMaterial() != Material.snow)
         {
       		return super.getDigSpeed(itemstack, block, metadata);
         }
         
         //for every charge level, efficiency increases by 4
-        return efficiencyOnProperMaterial + DMUtils.computeEfficiencyBonus(itemstack.getItemDamage());
+        return efficiencyOnProperMaterial + PowerItemUtils.computeEfficiencyBonus(itemstack.getItemDamage());
   
     }
 	
@@ -72,7 +78,7 @@ public class DMAxe extends ItemAxe implements IChargeable, IKeyBound
     {
     	if(key == Key.CHARGE)
     	{
-    		DMUtils.bumpChargeOnItem(itemStack);
+    		PowerItemUtils.bumpChargeOnItem(itemStack);
     	}
     }
